2nd round of feedback implemented, cleaup
