+++ search/search_service/proxy/base.py
@@ -84,20 +84,20 @@ class BaseProxy(metaclass=ABCMeta):
         pass
 
     @abstractmethod
-    def update_document_field(self, *,
-                              resource_key: str,
-                              resource_type: Resource,
-                              field: str,
-                              value: str = None,
-                              operation: str = 'add') -> str:
+    def update_document_by_key(self, *,
+                               resource_key: str,
+                               resource_type: Resource,
+                               field: str,
+                               value: str = None,
+                               operation: str = 'add') -> str:
         pass
 
     @abstractmethod
-    def delete_document_field(self, *,
-                              resource_key: str,
-                              resource_type: Resource,
-                              field: str,
-                              value: str = None) -> str:
+    def delete_document_by_key(self, *,
+                               resource_key: str,
+                               resource_type: Resource,
+                               field: str,
+                               value: str = None) -> str:
         pass
 
     @abstractmethod
