+++ search/search_service/proxy/elasticsearch.py
@@ -799,17 +799,17 @@ class ElasticsearchProxy(BaseProxy):
                filters: List[Filter]) -> SearchResponse:
         pass
 
-    def update_document_field(self, *,
-                              resource_key: str,
-                              resource_type: Resource,
-                              field: str,
-                              value: str = None,
-                              operation: str = 'add') -> str:
+    def update_document_by_key(self, *,
+                               resource_key: str,
+                               resource_type: Resource,
+                               field: str,
+                               value: str = None,
+                               operation: str = 'add') -> str:
         pass
 
-    def delete_document_field(self, *,
-                              resource_key: str,
-                              resource_type: Resource,
-                              field: str,
-                              value: str = None) -> str:
+    def delete_document_by_key(self, *,
+                               resource_key: str,
+                               resource_type: Resource,
+                               field: str,
+                               value: str = None) -> str:
         pass
